requirements.txt
Flask==2.0.2
google-cloud==0.34.0
google-cloud-core==2.2.1
google-cloud-pubsub==2.9.0
google-cloud-storage==1.43.0
gunicorn==20.1.0
numpy==1.19.5
pandas
torch==1.4.0
torchvision==0.5.0
urllib3==1.26.7
psutil-5.8.0
imblearn
torchsummary-1.5.1
tqdm-4.62.3